% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{AnnotateOntologiesSingleCellPortal}
\alias{AnnotateOntologiesSingleCellPortal}
\title{Add the Required meta data field for uploading the dataset to the Broad Single Cell Portal.  See (see https://singlecell.zendesk.com/hc/en-us/articles/360060609852-Required-metadata).  The values will be added in a "best-effort" way, but should be manually reviewed afterwards.}
\usage{
AnnotateOntologiesSingleCellPortal(
  seurat_object,
  maleCells = c(),
  femaleCells = c(),
  mixedCells = c(),
  biosampleField = "sample",
  donorField = "donor",
  species = "human",
  normalCells = c(),
  diseaseName = "AML",
  organName = "Blood",
  library_preparation_protocol = "10X 3' v3 Gene Expression",
  cellTypeField = "cellType"
)
}
\arguments{
\item{seurat_object}{Seurat Object,}

\item{maleCells}{id of male cells}

\item{femaleCells}{id of female cells}

\item{mixedCells}{id of mixed cells}

\item{biosampleField}{Existing field containing the biosample identifier}

\item{donorField}{Existing field containing the donor identifier}

\item{species}{Text field identifying the specie.  This will be matched against the taxonomy ontology}

\item{normalCells}{id of normal (non-diseased) cells}

\item{diseaseName}{Name of the disease affecting the non-normal cells. This will be matched against the MONDO ontology}

\item{organName}{Name of the organ under study. This will be matched against the UBERON ontology}

\item{library_preparation_protocol}{Name of library preparation protocol. This will be matched against the EFO ontology}

\item{cellTypeField}{Cell meta data field from the seurat_object containing the cell type names.  This will be matched against the cell ontology.}
}
\value{
a seurat object with the required metadata fields added (see https://singlecell.zendesk.com/hc/en-us/articles/360060609852-Required-metadata).  The use is expected to validate the results before uploading the data.
}
\description{
Add the Required meta data field for uploading the dataset to the Broad Single Cell Portal.  See (see https://singlecell.zendesk.com/hc/en-us/articles/360060609852-Required-metadata).  The values will be added in a "best-effort" way, but should be manually reviewed afterwards.
}
\examples{
\dontrun{
 seurat_object <- AnnotateOntologiesSingleCellPortal(seurat_object,femaleCells=colnames(seurat_object),
                                  biosampleField="sample",
                                  donorField="condition",
                                 species = "mouse",
                                 normalCells = seurat_object$condition == "NORMAL",
                                 diseaseName = "AML",
                                 organName = "Blood",
                                 library_preparation_protocol = "Flex",
                                 cellTypeField = "ClusterNameLevel1"  )  
}
}
