% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.r
\name{CollapseGenesetList}
\alias{CollapseGenesetList}
\title{Collapse a list of list of genesets into a single level list}
\usage{
CollapseGenesetList(genesetsLists, usePrefix = FALSE)
}
\arguments{
\item{genesetsLists}{List of list of genesets, as returned by LoadGeneSetsGMTs}

\item{usePrefix}{If TRUE, the top level list named a prefixed to the collapsed list names.  Default to FALSE to preserve the original geneset names.}
}
\value{
A single-level list genesets, ready to be used for pathway activity scoring.
}
\description{
Collapse a list of list of genesets into a single level list
}
\examples{
\dontrun{
genesetsLists = LoadGeneSetsGMTs("../genesets")
genesets <- CollapseGenesetList(genesetsLists)
genesets[[1]]
}
}
