% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.r
\name{ComputeAUCellPathwayActivity}
\alias{ComputeAUCellPathwayActivity}
\title{Run AUCell Pathway Activity scoring and threshold the results based on a null distribution obtained from a permuted gene matrix.}
\usage{
ComputeAUCellPathwayActivity(
  dataMat,
  genesets,
  p_val_thr = 0.05,
  minSize = 10,
  ncores = 1,
  ...
)
}
\arguments{
\item{dataMat}{A sparse scRNAseq gene expression matrix  (genes * cells)}

\item{genesets}{A list of genesets to compute the activity score}

\item{p_val_thr}{Pvalue threshold to use to threshold the results.  Pathway activity score values less significant than this p-value will be set to 0.}

\item{minSize}{Only use genesets having at least that many detected genes in the dataMatrix}

\item{ncores}{Use that number of cores when running AUCell}

\item{...}{Additional arguments passed on to UCell::AUCell_run}
}
\value{
A sparse pathway activity matrix (pathway * cells), which has been thresholded so that non significant scores are 0.
}
\description{
Run AUCell Pathway Activity scoring and threshold the results based on a null distribution obtained from a permuted gene matrix.
}
\examples{
\dontrun{
genesets = CollapseGenesetList(LoadGeneSetsGMTs("../genesets"))
geneExpMat = as(GetAssayData(object = seurat_object, assay = 'RNA', layer = "data"),"dgCMatrix")
pasMat = ComputeAUCellPathwayActivity(geneExpMat,genesets,ncores=5)
seurat_object[['PAS']] <- CreateAssayObject(data = pasMat)
}
}
