% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.r
\name{ComputeFCPseudobulk}
\alias{ComputeFCPseudobulk}
\title{Compute Fold Changes  between 2 conditions and multiple samples using a pseudobulk approach.  First, the assay is aggregated by sample and the data is normalized using a scaling factor corresponding to the average library depth of all the samples.  Then the FoldChange is compute over the data slot of the pseudobulk aggregated assay using Seurat standard "FoldChange" function.}
\usage{
ComputeFCPseudobulk(
  seurat_object,
  ident.1,
  ident.2,
  assay = "RNA",
  sampleId = "SampleId"
)
}
\arguments{
\item{seurat_object}{Seurat Object, with the condition of interest setup as "Ident"}

\item{ident.1}{Ident level of the First group.  Fold change is compute as (ident.1/ident.2)}

\item{ident.2}{Ident level of the Second group.  Fold change is compute as (ident.1/ident.2)}

\item{assay}{Assay name over which to compute fold changes.}

\item{sampleId}{Meta data field name containing the Id of the individual samples.}
}
\value{
A Fold change dataframe, with the assay feature names as rownames.
}
\description{
Compute Fold Changes  between 2 conditions and multiple samples using a pseudobulk approach.  First, the assay is aggregated by sample and the data is normalized using a scaling factor corresponding to the average library depth of all the samples.  Then the FoldChange is compute over the data slot of the pseudobulk aggregated assay using Seurat standard "FoldChange" function.
}
\examples{
\dontrun{
fcdf = ComputeFCPseudobulk(seurat_object,'condition1','condition2',assay='SCT',sampleId='orig.ident')
head(fcdf)
}
}
