% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportClusteringSingleCellPortal}
\alias{ExportClusteringSingleCellPortal}
\title{Export cell clustering to a format accepted by the Broad single cell Portal}
\usage{
ExportClusteringSingleCellPortal(
  seurat_object,
  file = "./clustering.txt",
  reduction = "UMAP",
  extraColumns = c("seurat_clusters")
)
}
\arguments{
\item{seurat_object}{Seurat Object,}

\item{file}{clustering file name}

\item{reduction}{the data reduction name to export (UMAP by default)}

\item{extraColumns}{Extra Columns to export alongside the UMAP}
}
\value{
nothing
}
\description{
Export cell clustering to a format accepted by the Broad single cell Portal
}
\examples{
\dontrun{
 ExportClusteringSingleCellPortal(seurat_object = seurat_object,
                                     file = "./metadata.txt",
                                     reduction = "umap")
}
}
