% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportLabelCoordinatesSingleCellPortal}
\alias{ExportLabelCoordinatesSingleCellPortal}
\title{Export coordinate label files to a format accepted by the Broad single cell Portal}
\usage{
ExportLabelCoordinatesSingleCellPortal(
  seurat_object,
  file = "./labels.txt",
  reduction = "UMAP",
  label = "seurat_clusters"
)
}
\arguments{
\item{seurat_object}{Seurat Object,}

\item{file}{meta data file name}

\item{reduction}{The reduction to base the coordinate on}

\item{label}{The Group label column to use}
}
\value{
nothing
}
\description{
Export coordinate label files to a format accepted by the Broad single cell Portal
}
\examples{
\dontrun{
 ExportLabelCoordinatesSingleCellPortal(seurat_object = seurat_object,
                                     file = "./labels.txt",
                                     reduction = "UMAP",
                                     label = "cellType"
)
}
}
