% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportMatrixSingleCellPortal}
\alias{ExportMatrixSingleCellPortal}
\title{Export a matrix to the mtx format accepted by the Broad single cell Portal}
\usage{
ExportMatrixSingleCellPortal(
  seurat_object,
  folder = "./processed",
  assay = "RNA",
  layer = "data"
)
}
\arguments{
\item{seurat_object}{Seurat Object}

\item{folder}{Folder where to export the matrix}

\item{assay}{Assay name to export}

\item{layer}{Layer to export (data) by default}
}
\value{
nothing
}
\description{
Export a matrix to the mtx format accepted by the Broad single cell Portal
}
\examples{
\dontrun{
ExportMatrixSingleCellPortal(seurat_object = seurat_object,
                                     folder = "./processed"
                                     assay = "RNA",
                                     layer = "data")
}
}
