% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportMetadataSingleCellPortal}
\alias{ExportMetadataSingleCellPortal}
\title{Export cell metadata to a format accepted by the Broad single cell Portal}
\usage{
ExportMetadataSingleCellPortal(
  seurat_object,
  file = "./metadata.txt",
  columns = colnames(seurat_object@meta.data)
)
}
\arguments{
\item{seurat_object}{Seurat Object,}

\item{file}{meta data file name}

\item{columns}{Columns to export (default to all metadata columns)}
}
\value{
nothing
}
\description{
Export cell metadata to a format accepted by the Broad single cell Portal
}
\examples{
\dontrun{
 ExportMetadataSingleCellPortal(seurat_object = seurat_object,
                                     file = "./metadata.txt")
}
}
