% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportPseudoBulkAssay}
\alias{ExportPseudoBulkAssay}
\title{Export a pseudobulk assay to an Excel workbook.  2 sheets will be created, one with the assay counts data (PseudoBulk_RawCounts) and another one with the normalized log2CPM data (PseudoBulk_log2CPMNorm)}
\usage{
ExportPseudoBulkAssay(
  seurat_object,
  file,
  assay = "RNA",
  features = NULL,
  group.by = "ident",
  tableStyle = "TableStyleLight9"
)
}
\arguments{
\item{seurat_object}{Seurat Object, with the condition of interest setup as "Ident"}

\item{file}{Excel file name (and path)}

\item{assay}{Assay name to export}

\item{features}{Subset of features to export (or NULL to export all features)}

\item{group.by}{Metadata fields over which to compute the pseudobulk assay.  Set to "ident" by default.}

\item{tableStyle}{Style of the data table in the excel workbook}
}
\value{
a seurat object representing the pseudo_bulk assay
}
\description{
Export a pseudobulk assay to an Excel workbook.  2 sheets will be created, one with the assay counts data (PseudoBulk_RawCounts) and another one with the normalized log2CPM data (PseudoBulk_log2CPMNorm)
}
\examples{
\dontrun{
pb_seurat <- ExportPseudoBulkAssay(seurat_object = seurat_object,
                                     assay = "RNA",
                                     group.by = c("SampleId", "Condition"),
                                     file = "./path/to/export/file.xlsx")
}
}
