% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportTables}
\alias{ExportTables}
\title{Export a list of tables to 1 excel file}
\usage{
ExportTables(
  tableList,
  file,
  colNames = TRUE,
  rowNames = FALSE,
  freezeFirstRow = TRUE,
  freezeFirstCol = TRUE,
  tableStyle = "TableStyleLight9"
)
}
\arguments{
\item{tableList}{Named List of tables to export.  The name will be used as the sheet name in the output Excel file}

\item{file}{Excel file name (and path)}

\item{colNames}{Export column names (TRUE / FALSE)}

\item{rowNames}{Export row names (TRUE / FALSE)}

\item{freezeFirstRow}{Set to TRUE to freeze the first row.}

\item{freezeFirstCol}{Set to TRUE to freeze the first column}

\item{tableStyle}{Style of the data table in the excel workbook}
}
\description{
Export a list of tables to 1 excel file
}
\examples{
\dontrun{
ExportTables(tableList = List("table1" = df1,
                              "table2" = df2),
             file = "./path/to/export/file.xlsx")
}
}
