% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.r
\name{ExportToH5AD}
\alias{ExportToH5AD}
\title{Export a seurat object to H5AD}
\usage{
ExportToH5AD(
  seurat_object,
  file = "adata.h5ad",
  assay = "RNA",
  dimreducs = Reductions(seurat_object)
)
}
\arguments{
\item{seurat_object}{Seurat Object,}

\item{file}{h5ad file name}

\item{assay}{The assay to export}

\item{dimreducs}{The list of data reduction to export}
}
\value{
nothing
}
\description{
Export a seurat object to H5AD
}
\examples{
\dontrun{
 ExportToH5AD(seurat_object = seurat_object,
                                     file = "my_data.h5ad",
                                     reduction = c("umap")
)
}
}
