% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.r
\name{find_de_MAST_RE}
\alias{find_de_MAST_RE}
\title{Perform Differential Gene Expression using MAST 
Adapted from https://www.sc-best-practices.org/conditions/differential_gene_expression.html#single-cell-specific}
\usage{
find_de_MAST_RE(
  sca,
  formulaToUse,
  conditionOfInterest,
  minProportionOfCell = 0.1,
  libSizeVar = "ngeneson",
  filterFDRThr = 0.05,
  mixed = T
)
}
\arguments{
\item{sca}{SinglecellAssay object over which to perform the DGE}

\item{formulaToUse}{DE Formula to apply}

\item{conditionOfInterest}{Contrast for which the p-value and fold change should be computed}

\item{minProportionOfCell}{Minimum global proportion of the cells expression a given gene.  Genes expressed in less cells than that are not tested for DGE.}

\item{libSizeVar}{Variable name representing the number of genes expressed per cells (as used in the formulaToUse)}

\item{filterFDRThr}{Filter output results based on this FDR threshold}

\item{mixed}{Set to true if the formula represents a Mixed model.  Set to false otherwise.}
}
\value{
A data frame with DGE results.
}
\description{
Perform Differential Gene Expression using MAST 
Adapted from https://www.sc-best-practices.org/conditions/differential_gene_expression.html#single-cell-specific
}
