% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.r
\name{LoadGeneSetsGMTs}
\alias{LoadGeneSetsGMTs}
\title{Load Genesets files in GMT format from a folder}
\usage{
LoadGeneSetsGMTs(folder = ".")
}
\arguments{
\item{folder}{Path where the GMT files are stored (directory)}
}
\value{
A list of list of geneset, the top level list is a named list of GMT files and the lower level list is a named list of the genesets within this file.
}
\description{
Load Genesets files in GMT format from a folder
}
\examples{
\dontrun{
genesets = LoadGeneSetsGMTs("../genesets")
names(genesets)
genesets[[1]]
}
}
