% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read10XMoleculeInfo.r
\name{Read10XMoleculeInfo}
\alias{Read10XMoleculeInfo}
\title{Read the 10X Molecule_info.h5 file and outputs separated EXONS/INTRONS count matrices along with basic cell and feature metadata}
\usage{
Read10XMoleculeInfo(
  moleculeFile = "molecule_info.h5",
  featureType = "Gene Expression",
  filtered = TRUE
)
}
\arguments{
\item{moleculeFile}{Path to the molecule_info.h5 file.  Defaults to the molecule_info.h5 file in the current directory.}

\item{featureType}{Feature type to extract.  Defaults to 'Gene Expression'}

\item{filtered}{True to return the Filtered matrix (only cells).  False to return the full matrix of all barcodes.  Defaults to True.}
}
\value{
A list with the following 4 elements : EXONS , INTRONS , cells.meta.data, features.meta.data
}
\description{
Read the 10X Molecule_info.h5 file and outputs separated EXONS/INTRONS count matrices along with basic cell and feature metadata
}
\examples{
\dontrun{
counts = Read10XMoleculeInfo(moleculeFile="cellranger/sample/outs/molecule_info.h5")
counts$EXONS # The exon count matrix
counts$INTRONS # The intron count matrix
counts$cells.meta.data # data.frame of cells meta data
counts$features.meta.data # data.frame of features meta data
seurat_object <- CreateSeuratObject(counts$EXONS+counts$INTRONS,
meta.data=counts$cells.meta.data,
assay = "RNA")
seurat_object[['INTRONS']] <- CreateAssayObject(counts = counts$INTRONS )
seurat_object[['EXONS']] <- CreateAssayObject(counts = counts$EXONS )
seurat_object[['RNA']] <- AddMetaData(seurat_object[['INTRONS']], counts$features.meta.data)
seurat_object[['INTRONS']] <- AddMetaData(seurat_object[['INTRONS']], counts$features.meta.data)
seurat_object[['EXONS']] <- AddMetaData(seurat_object[['EXONS']], counts$features.meta.data)
}
}
