% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.r
\name{RunDGE_MAST_GLMModel}
\alias{RunDGE_MAST_GLMModel}
\title{Perform Differential Gene Expression using MAST using a GLM model c The fold changes are however computed using a pseudo-bulk approach (see ComputeFCPseudobulk)}
\usage{
RunDGE_MAST_GLMModel(
  seurat_object,
  ident.1,
  ident.2,
  formula = "~ ngeneson + group",
  sampleId = "sampleId",
  assay = "RNA",
  identVar = "group",
  librarysize_var = "ngeneson",
  minProportionOfCell = 0.05,
  filterFDRThr = 2
)
}
\arguments{
\item{seurat_object}{Seurat Object, with the condition of interest setup as "Ident"}

\item{ident.1}{Ident level of the First group.  Fold change is compute as (ident.1/ident.2)}

\item{ident.2}{Ident level of the Second group.  Fold change is compute as (ident.1/ident.2)}

\item{formula}{DE Formula to apply}

\item{sampleId}{Meta data field name containing the Id of the individual samples.}

\item{assay}{Assay name over which to compute fold changes.}

\item{identVar}{Name of the Ident var within the provided formula}

\item{minProportionOfCell}{Minimum global proportion of the cells expression a given gene.  Genes expressed in less cells than that in one of the 2 groups are not tested for DGE.}

\item{filterFDRThr}{Filter output results based on this FDR threshold}

\item{libSizeVar}{Variable name representing the number of genes expressed per cells (as used in the formulaToUse)}
}
\value{
A data frame with DGE results.
}
\description{
Perform Differential Gene Expression using MAST using a GLM model c The fold changes are however computed using a pseudo-bulk approach (see ComputeFCPseudobulk)
}
