% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.r
\name{RunFGSEAMultilevelBatch}
\alias{RunFGSEAMultilevelBatch}
\title{Run FGSEAMultilevel in batch mode, for each rankedGeneLists and for each type of Genesets}
\usage{
RunFGSEAMultilevelBatch(
  genesetLists,
  rankedGeneLists,
  removeGenePattern = NULL,
  geneListsGroupedBy = "RankedGeneList",
  ...
)
}
\arguments{
\item{genesetLists}{A lists of list of genesets, as returned by LoadGeneSetsGMTs}

\item{rankedGeneLists}{A list of ranked gene lists, as returned by CreateRanking}

\item{removeGenePattern}{Optional pattern to remove unwanted genes from the ranking before running FGSEA.  Could be used to remove Ribosomal or mitochondrial genes, for instance.}

\item{geneListsGroupedBy}{Column name in the output data frame used to identify the different ranked gene lists}

\item{...}{Additional arguments passed on to fgseaMultilevel}
}
\value{
A dataframe containing the FGSEA results.  The leadingEdge column is collapsed into a string delimited by ;
}
\description{
Run FGSEAMultilevel in batch mode, for each rankedGeneLists and for each type of Genesets
}
\examples{
\dontrun{
genesets = LoadGeneSetsGMTs("../genesets")
rankings = CreateRanking(myDEResultsDF,pvalColumn='p_val',log2FCColumn="avg_Log2FC",geneColumn='gene',groupByColumn="cluster")
fgseaResults <- RunFGSEAMultilevelBatch(genesets,rankings,removeGenePattern="RPS|RPL|MT-")
head(fgseaResults)
}
}
