% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.r
\name{split_data.matrix}
\alias{split_data.matrix}
\title{Split data matrix into smaller sub-matrices ('chunks')
Note : Code copied from https://rdrr.io/github/carmonalab/UCell/man/split_data.matrix.html}
\usage{
split_data.matrix(matrix, chunk.size = 100)
}
\arguments{
\item{matrix}{Input data matrix}

\item{chunk.size}{How many cells to include in each sub-matrix}
}
\value{
A list of sub-matrices, each with size {n_features x chunk_size}
}
\description{
Split data matrix into smaller sub-matrices ('chunks')
Note : Code copied from https://rdrr.io/github/carmonalab/UCell/man/split_data.matrix.html
}
